feat: implement loading challenge and project objects
Refactored config loading to use a loader object bound to a Project
instance.

Implemented Project and Challenge objects to hold its config data and
act as a context for other objects to be constructed off of (the
ChallengeLoader uses a Project to create a Challenge).

Currently, a Project will always create a Docker client. This should
probably be refactored to be loaded on-demand.