feat: initial container building implementation
Add ContainerManager and Container objects for extracting and building
containers for challenges. Tag images based on the hash of the content
in the build context for each built container.

Tests: Currently, only the build context hashing and the creation of
Container objects is being tested. Testing / mocking needs to be set up
for the logic for checking whether or not the containers are up-to-date
or need building.