feat(assets): use explicit manifest containing file list
Since certain OS's / tools (macOS) can create files on the disk, we
can't use the filesystem as a source of truth for what assets are in the
asset manager. Instead, use an explicit manifest that is persisted as a
JSON file to track what files are in the manager.

Also, refactor the implementation to stop checking for extra files
whenever possible, and instead only do that when we sync the manifest to
dist.

The manifest loading and storing still needs tests, as well as
extraneous file removal and erroring on missing files.

The two tests added are unrelated to the functionality added in this
commit.