refactor(config): return the parsed config from the generator
To be able to support steps in the config validation which modify the
config, we need to extract the loaded config at the end of all
validation. So, reimplement validation by having a parse step which
yields both errors and the fully loaded config; keep the old API by
wrapping this function and checking if the first value yielded is a
ValidationError.

This commit also adds an extra test for schema validation errors, and
adds some extra assertions in existing tests.