feat: project config schema and docker image tagging improvement
Add a schema for the project-level rcds.yaml, and also improve the
generation of image tag names for built docker containers by using a
configurable Jinja template. This also adds a concept of an id to a
challenge (which is currently not checked for uniqueness across the
project - this needs to be implemented soon). The id is by default
derived from the name of the directory the challenge is located in, but
can be overridden in the challenge.yml.