test(config): fix attempt to load wrong challenge directory
The test for the loader throwing when given an invalid challenge config
was previously using the wrong project directory (which would then cause
the loader to throw because it couldn't open the file, thus passing the
test). Fix this to actually load the correct challenge, and test that
the error thrown is actually a ValidationError.