fix(docker): fix ignore file parsing
pathspec's behavior with negated patterns is different than that of
docker (and its own behavior with non-negated patterns) in that patterns
ending with `/` will match files in subdirectories, but not a file with
the same name, and pattens not ending in `/` will only match files, but
not files in subdirectories. For example, the pattern `!/a` will exclude
`a`, but not `a/b`, and the pattern `!/a/` will exclude `a/b`, but not
`a`. Since docker treats these interchangeably, add some code to
automatically insert the corresponding ignore rule into the rules list
if a negated pattern is detected (insert `!/a/` if `!/a` is detected and
vice versa).