fix(assets): make file / link handling mutually exclusive
If a file being written to the asset cache is detected to be an existing
file (and thus will be linked), skip checks for if the file is in-memory
data (and copying if it is). This avoids issues in pathological cases
where an object which is a subclass of both `Path` and a bytes or
BinaryIO like object is passed.

This commit was in response to mypy 0.800 no longer narrowing the type
to solely `BinaroIO` in the corresponding branch.