ci: stop using matrix for single-version jobs
For jobs which only run on a single python version (everything except
test), do not use the matrix so the version is not in the job name. This
way the python version can be updated without needing to change the
required status checks list on the GitHub branch protection rule.